refactor: input
